Correctly gather package-level version numbers from xsi files
The previous dist failed to index because FLTK::Widget's version number
seemed to decrease from the previous stable. The metadata was bad because
xs/Cursor.xsi has a FLTK::Widget block which was overriding the actual
version in xs/Widget.xsi. Now, we store the current package as soon as we
find the version number.